+++ tasks/task_178e.py
